Add unchecked sectors
In some cases it's helpful to suppress the checks on a per-sector level.
Since sectors can be renumbered, and since extraneous lumps are often
removed, a special thing with number 1234 is used to indicate that the
containing sector should be exempt from checks. Such sectors will be
referred to as "unchecked sectors".

Items that are directly related to the uncheked sectors will also be
uncheked. Currently the uncheked item types are:
  * sectors
  * sidedefs
  * things

So far the following checks have been implemented:
  * Unknown things
  * Missing textures
  * Medusa textures

The above was driven by the needs of the Freedoom project - that
project makes use of fake floors and fake ceilings, which results in
false missing texture positives. It also uses textures that are known
to work which have false Medusa positives.

Currently the user indicates that a sector is to be uncheked by adding
thing #1234 to it, which appears as a pink "?" in Eureka. A custom
sprite would be nice, but does not exist yet.